refactor ExchangeRates to components
