+++ src/components/ExchangeRates.js
@@ -1,0 +1,26 @@
+import React from 'react'
+import { gql, useQuery } from '@apollo/client'
+
+const EXCHANGE_RATES = gql`
+  {
+    rates(currency: "USD") {
+      currency
+      rate
+    }
+  }
+`
+function ExchangeRates() {
+  const { loading, error, data } = useQuery(EXCHANGE_RATES)
+
+  if (loading) return <p>Loading...</p>
+  if (error) return <p>error...</p>
+  return data.rates.map(({ currency, rate }) => (
+    <div key={currency}>
+      <p>
+        {currency}: {rate}
+      </p>
+    </div>
+  ))
+}
+
+export default ExchangeRates
